Factorisation et débug test avancer

+++ MarsRover/Rover.cs
@@ -3,10 +3,10 @@
 public class Rover
 {
     public PointCardinal Orientation { get; }
-    public uint Latitude { get; }
-    public uint Longitude { get; }
+    public int Latitude { get; }
+    public int Longitude { get; }
 
-    public Rover(PointCardinal orientationDépart, uint latitude = 0, uint longitude = 0)
+    public Rover(PointCardinal orientationDépart, int latitude = 0, int longitude = 0)
     {
         Orientation = orientationDépart;
         Latitude = latitude;
@@ -17,7 +17,7 @@ public class Rover
     public Rover TournerAGauche() => new(Orientation.SuivantAntihoraire);
 
     public Rover Avancer() => new (Orientation, 
-        (uint) (Latitude + Orientation.VecteurLatitude),
-        (uint) (Latitude + Orientation.VecteurLongitude)
+        Latitude + Orientation.VecteurLatitude,
+        Latitude + Orientation.VecteurLongitude // BUG 
     );
 }
